# Conflicts:
#	src/main/java/com/example/datenbank/controller/SchadenController.java

+++ src/main/java/com/example/datenbank/service/SchadenCRUD.java
@@ -18,7 +18,7 @@ public class SchadenCRUD {
         try {
             connection.getDBConnection();
             PreparedStatement st = connection.getCon().prepareStatement(sql);
-            st.setBigDecimal(1, schaden.getHoehe());
+            st.setInt(1, schaden.getHoehe());
             st.setString(2, schaden.getBeschreibung());
             st.execute();
             st.close();
@@ -36,7 +36,7 @@ public class SchadenCRUD {
             ResultSet rs = stmt.executeQuery();
             Schaden schaden;
             while (rs.next()) {
-                schaden = new Schaden(rs.getInt("Schaden_ID"), rs.getBigDecimal("Hoehe"), rs.getString("Beschreibung"));
+                schaden = new Schaden(rs.getInt("Schaden_ID"), rs.getInt("Hoehe"), rs.getString("Beschreibung"));
                 list.add(schaden);
             }
             rs.close();
@@ -55,7 +55,7 @@ public class SchadenCRUD {
                     "UPDATE Schaden " +
                             "SET Hoehe = ?, Beschreibung = ? " +
                             "WHERE Schaden_ID = ?");
-            stmt.setBigDecimal(1, schaden.getHoehe());
+            stmt.setInt(1, schaden.getHoehe());
             stmt.setString(2, schaden.getBeschreibung());
             stmt.setInt(3, schaden.getSchadenID());
             stmt.execute();
